Added basic login files
Current issues: cannot properly access database without error.

Currently only testing. The index.php automatically sends the user to login, so that we don't have to append the file name to the end of the URL. The testHome.php as the home page that logging in sends the user on a successful login.